Emitting 'show' and 'hide' events when they're a bit more usable.
If I want to render a dialog and focus the element I can't focus the element before it's attached to the dom and visible. So triggering show before it's attached to the dom makes that hard.

Also, 'hide' would be triggered at what seems like the wrong time if you set a delay.